Add files via upload
This project is a basic API server created as part of the HNG Backend Track Stage One task. The server utilizes PHP to fetch weather information based on the client's IP address and provides a greeting message with temperature details. The weather data is fetched from the WeatherAPI.com service.